Implement reply delete action

+++ Web/ForumNet.Web/Controllers/PostsController.cs
@@ -178,17 +178,17 @@
         [ActionName("Delete")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var post = await this.postsService.GetByIdAsync<PostsDeleteDetailsViewModel>(id);
+            var post = await this.postsService.GetByIdAsync<PostsDeleteConfirmedViewModel>(id);
             if (post == null)
             {
                 return this.NotFound();
             }
 
-            //var currentUserId = this.User.GetId();
-            //if (post.AuthorId != currentUserId && !this.User.IsAdministrator())
-            //{
-            //    return this.Unauthorized();
-            //}
+            var currentUserId = this.User.GetId();
+            if (post.AuthorId != currentUserId && !this.User.IsAdministrator())
+            {
+                return this.Unauthorized();
+            }
 
             await this.postsService.DeleteAsync(id);
